fix: perform playtime division in i64 before i32 cast for IDs 161-164
IDs 161-164 (elapsed/remaining playtime minutes and seconds) cast i64
timer values to i32 before dividing by 60000/1000. For songs over ~35.8
minutes, the i64 value exceeds i32::MAX and wraps to negative, producing
wrong display values. Move the as-i32 cast after division/modulo so
arithmetic stays in i64 space. Applied to both PlayRenderContext and
PlayMouseContext.

+++ crates/rubato-play/src/bms_player/skin_context.rs
@@ -245,18 +245,18 @@ impl rubato_types::skin_render_context::SkinRenderContext for PlayRenderContext<
             92 => self.main_bpm as i32,
             160 => self.now_bpm as i32,
             // Elapsed playtime from TIMER_PLAY (Java: timer.getNowTime(TIMER_PLAY))
-            161 => (self.timer.now_time_for_id(TIMER_PLAY) as i32) / 60000,
-            162 => ((self.timer.now_time_for_id(TIMER_PLAY) as i32) / 1000) % 60,
+            161 => (self.timer.now_time_for_id(TIMER_PLAY) / 60000) as i32,
+            162 => ((self.timer.now_time_for_id(TIMER_PLAY) / 1000) % 60) as i32,
             // Remaining playtime (Java: max(playtime - elapsed + 1000, 0))
             163 => {
-                let elapsed = self.timer.now_time_for_id(TIMER_PLAY) as i32;
-                let remaining = (self.playtime as i32 - elapsed + 1000).max(0);
-                remaining / 60000
+                let elapsed = self.timer.now_time_for_id(TIMER_PLAY);
+                let remaining = (self.playtime - elapsed + 1000).max(0);
+                (remaining / 60000) as i32
             }
             164 => {
-                let elapsed = self.timer.now_time_for_id(TIMER_PLAY) as i32;
-                let remaining = (self.playtime as i32 - elapsed + 1000).max(0);
-                (remaining / 1000) % 60
+                let elapsed = self.timer.now_time_for_id(TIMER_PLAY);
+                let remaining = (self.playtime - elapsed + 1000).max(0);
+                (remaining / 1000 % 60) as i32
             }
             // Scroll duration from LaneRenderer (Java: getCurrentDuration())
             312 => self.current_duration,
@@ -706,18 +706,18 @@ impl rubato_types::skin_render_context::SkinRenderContext for PlayMouseContext<'
                 .as_ref()
                 .map_or(0, |lr| lr.now_bpm() as i32),
             // Elapsed playtime from TIMER_PLAY
-            161 => (self.timer.now_time_for_id(TIMER_PLAY) as i32) / 60000,
-            162 => ((self.timer.now_time_for_id(TIMER_PLAY) as i32) / 1000) % 60,
+            161 => (self.timer.now_time_for_id(TIMER_PLAY) / 60000) as i32,
+            162 => ((self.timer.now_time_for_id(TIMER_PLAY) / 1000) % 60) as i32,
             // Remaining playtime
             163 => {
-                let elapsed = self.timer.now_time_for_id(TIMER_PLAY) as i32;
-                let remaining = (self.player.playtime as i32 - elapsed + 1000).max(0);
-                remaining / 60000
+                let elapsed = self.timer.now_time_for_id(TIMER_PLAY);
+                let remaining = (self.player.playtime - elapsed + 1000).max(0);
+                (remaining / 60000) as i32
             }
             164 => {
-                let elapsed = self.timer.now_time_for_id(TIMER_PLAY) as i32;
-                let remaining = (self.player.playtime as i32 - elapsed + 1000).max(0);
-                (remaining / 1000) % 60
+                let elapsed = self.timer.now_time_for_id(TIMER_PLAY);
+                let remaining = (self.player.playtime - elapsed + 1000).max(0);
+                (remaining / 1000 % 60) as i32
             }
             // Scroll duration from LaneRenderer (Java: getCurrentDuration())
             312 => self
